Signals kill children, and JAR is run in bg
When sh gets INT or TERM signals, we want to send SIGTERM to the
java process.
Experimentation shows that bash only honors signal handlers if the
java process is backgrounded.

Quirks: sometimes this script's `kill 0` kills the script itself,
which causes it to exit with status 143 and print 'Terminated'
to STDERR, but other times it reaches the end of the signal handler
first, which exits with status 1.