Added capability for providedDependencies to be "internal".
The best use case for this is nCipher's JCE provider. It is signed
and thus must be "provided". However, it is not a publicly accessible
jar. The best approach seems to be a module called
external.ncipher.ncipherkm, which has a jars = {nCipherKM.jar}
entry in it's bakefile.

The walk strategy is modified to preclude modules which match
non-external dependencies listed as providedDependencies.

The new jars from internal dependencies are added to the
manifest's Class-Path.